1672: refactor solution
